+++ src/tests/test_parse.py
@@ -203,6 +203,7 @@ def complete_with_defaults(mapping):
         "creator": None,
         "section": None,
         "persist": False,
+        "persist_replace": False,
         "no_index": False,
         "append": False,
         "connection_arguments": None,