+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.6dev
 
 * [Feature] Support for printing capture variables using `=<<` syntax (by [@jorisroovers](https://github.com/jorisroovers))
+* [Feature] Adds `--persist-replace` argument to replace existing tables when persisting data frames (#440)
 
 ## 0.7.5 (2023-05-24)